Big update. 1. Added a full checkAdmin functionality. Deployed it wherever /todo add admin check was added. In order for it to work a protected UserService.java instance ref was needed to be added to AbstractController.java 2.Added a protection against session high jacking method. Deployed it with check admin func and in User Controller. 3. Moved all checks that use HTTP request or session to the controller. 3.Code refactoring. 4. User-Order relationship One-to-Many 5. Minor bug fixes and reformats. 6. Removed a few unused imports.

+++ src/main/java/com/example/emag/controller/CategoryController.java
@@ -4,14 +4,12 @@ import com.example.emag.model.dto.category.CategoryDTO;
 import com.example.emag.model.dto.category.CategoryRequestDTO;
 import com.example.emag.model.dto.category.CategoryWithSubcategoryDTO;
 import com.example.emag.model.dto.product.ProductDTO;
-import com.example.emag.model.entities.Category;
-import com.example.emag.model.exceptions.NotFoundException;
-import com.example.emag.model.repositories.CategoryRepository;
 import com.example.emag.service.CategoryService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -22,20 +20,27 @@ public class CategoryController extends AbstractController {
     @Autowired
     private CategoryService categoryService;
     @PostMapping("/categories")
-    public CategoryDTO addCategory(@RequestBody CategoryRequestDTO category){
-        //todo check if admin
+    public CategoryDTO addCategory(@RequestBody CategoryRequestDTO category, HttpServletRequest req){
+        checkIfLogged(req);
+        checkIpWithSessionIp(req);
+        checkIfAdmin(req);
         return categoryService.save(category);
     }
 
     @DeleteMapping(value = "/categories/{id}", headers = "password=dve")
-    public CategoryDTO deleteCategory(@PathVariable long id){
-        // todo check if admin
+    public CategoryDTO deleteCategory(@PathVariable long id, HttpServletRequest req){
+        checkIfLogged(req);
+        checkIpWithSessionIp(req);
+        checkIfAdmin(req);
         return categoryService.remove(id);
     }
 
     @PutMapping("categories/{id}")
-    public CategoryWithSubcategoryDTO editCategory(@PathVariable long id, @RequestBody CategoryRequestDTO dto){
-        //todo check if admin
+    public CategoryWithSubcategoryDTO editCategory(@PathVariable long id,
+                                                   @RequestBody CategoryRequestDTO dto, HttpServletRequest req){
+        checkIfLogged(req);
+        checkIpWithSessionIp(req);
+        checkIfAdmin(req);
         return categoryService.edit(id, dto);
     }
 
@@ -49,8 +54,11 @@ public class CategoryController extends AbstractController {
         return categoryService.getAllCategories();
     }
     @PostMapping("categories/{cid}/discount/{did}")
-    public List<ProductDTO> setDiscountForCategory(@PathVariable long cid, @PathVariable long did){
-        //todo check if admin
+    public List<ProductDTO> setDiscountForCategory(@PathVariable long cid,
+                                                   @PathVariable long did, HttpServletRequest req){
+        checkIfLogged(req);
+        checkIpWithSessionIp(req);
+        checkIfAdmin(req);
         return categoryService.setDiscountForCategory(cid, did);
     }
 }
